spec: Add dependences for crash utility
crash utility on Fedora Linux 38 merged the patch for support of
kernel module memory layout change on Linux kernel 6.4 at 8.0.3-1:

    * Tue Jun 27 2023 Lianbo Jiang <lijiang@redhat.com> - 8.0.3-1
    - Rebase to upstream crash 8.0.3
    - Support module memory layout change on Linux 6.4

The previous commit was for crash trace to work well with this change.

To built crash trace now, we need new macros and data structures
introduced at the change. Hence, add BuildRequires dependency to
require crash-devel of the version 8.0.3-1 or newer.

In addition, add Requires dependency to require crash utility of the
version 8.0.3-1 or newer. This is because the change above gives crash
trace not only built-time but also runtime dependency, making crash
trace not work well with the crash utility older than 8.0.3-1.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@fujitsu.com>